requirements.txt
appdirs==1.4.0
Django==1.8.3
funcsigs==1.0.2
httplib2==0.10.3
httpretty==0.8.14
linecache2==1.0.0
mock==2.0.0
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
packaging==16.8
pbr==1.10.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
PyJWT==1.4.2
pyparsing==2.1.10
python-openid==2.2.5
python-social-auth==0.2.12
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.7.0
rsa==3.4.2
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
traceback2==1.4.0
unittest2==1.1.0
